Problem 2 Changes
Mostly changes to problem 2. Included comments for the problem. Also
changed some of the cat() functions to paste() because the cat()
function, for some reason, was not able to be plugged into the objects
m.significances and d.significances. Also modified problem 1 to include
the square root of n in the equations (as on facebook page). NOTE TO
SELF: Still need a way for the table output from problem 2 to line up
better when the number of stars are different.